fix: move chart props to interfaces file.
