+++ src/components/analytics/technicals/charts/CMFChart.tsx
@@ -1,11 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
-import { fetchSingleChart } from "../../../utilities/api";
+import { fetchSingleChart } from "../../../../utilities/api";
+import { CMFChartProps } from "../../../../types/interfaces";
 
-interface VolumeChartProps {
-  symbol: string;
-}
-
-const VolumeChart: React.FC<VolumeChartProps> = ({ symbol }) => {
+const CMFChart: React.FC<CMFChartProps> = ({ symbol }) => {
   const [chartHtml, setChartHtml] = useState<string>("");
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
@@ -14,10 +11,7 @@ const VolumeChart: React.FC<VolumeChartProps> = ({ symbol }) => {
   useEffect(() => {
     const loadChart = async () => {
       try {
-        const response = await fetchSingleChart(
-          "on_balance_volume_chart",
-          symbol
-        );
+        const response = await fetchSingleChart("adx_plot", symbol);
         setChartHtml(response); // Save the HTML content
       } catch {
         setError("Failed to load chart.");
@@ -59,4 +53,4 @@ const VolumeChart: React.FC<VolumeChartProps> = ({ symbol }) => {
   );
 };
 
-export default VolumeChart;
+export default CMFChart;
